refactor(role_service): move mappers to application/mapper
