+++ src/main/java/com/renewsim/backend/role_service/application/service/RoleApplicationService.java
@@ -17,7 +17,7 @@ import com.renewsim.backend.shared.domain.vo.RoleName;
 import com.renewsim.backend.role_service.domain.policy.RolePolicy;
 import com.renewsim.backend.role_service.domain.service.RoleDomainService;
 import com.renewsim.backend.role_service.dto.RoleDTO;
-import com.renewsim.backend.role_service.infrastructure.mapper.RoleDtoMapper;
+import com.renewsim.backend.role_service.application.mapper.RoleDtoMapper;
 
 @Service
 @Transactional
